move show acceptance test to show_resources file
